% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGAselID.R
\name{Crossover}
\alias{Crossover}
\title{Crossover}
\usage{
Crossover(c1, c2, chrConf)
}
\arguments{
\item{c1}{Set of chromosomes.}

\item{c2}{Set of chromosomes.}

\item{chrConf}{Configuration of chromosomes returned by splitChromosomes().}
}
\description{
Two-point crossover operator.
}
\examples{
\dontrun{
library(ALL)
data(ALL)

demoALL<-ALL[1:12,1:8]
set.seed(1357)
population02<-InitialPopulation(demoALL, 2, 4, FALSE)
chrConf02<-splitChromosomes(demoALL, 2)
chrConf02
population02[1:2,]
Crossover(population02[1,], population02[2,], chrConf02)
 }
}

