% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGAselID.R
\name{Individuals}
\alias{Individuals}
\title{Individuals}
\usage{
Individuals(population)
}
\arguments{
\item{population}{Population of haploid genotypes.}
}
\description{
Generates individuals with diploid genotypes.
}
\examples{
\dontrun{
library(ALL)
data(ALL)

demoALL<-ALL[1:12,1:8]

population02<-InitialPopulation(demoALL, 20, 4, FALSE)
individuals02<-Individuals(population02)
 }
}

