% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGAselID.R
\name{InitialPopulation}
\alias{InitialPopulation}
\title{InitialPopulation}
\usage{
InitialPopulation(x, populationSize, startGenes,
  EveryGeneInInitialPopulation = TRUE)
}
\arguments{
\item{x}{Dataset in ExpressionSet format.}

\item{populationSize}{Number of genotypes in initial population.}

\item{startGenes}{Genes in the genotypes at initialization.}

\item{EveryGeneInInitialPopulation}{Request for every gene to be present in the initial population. The default value is TRUE.}
}
\description{
Generates an initial randomly generated population of haploid genotypes.
}
\examples{
\dontrun{
library(ALL)
data(ALL)

demoALL<-ALL[1:12,1:8]

population01<-InitialPopulation(demoALL, 4, 4)
population02<-InitialPopulation(demoALL, 20, 4, FALSE)
 }
}

