% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGAselID.R
\name{PlotGenAlg}
\alias{PlotGenAlg}
\title{PlotGenAlg}
\usage{
PlotGenAlg(DGenes, dGenes, maxEval, meanEval)
}
\arguments{
\item{DGenes}{Occurences of genes as dominant.}

\item{dGenes}{Occurences of genes as recessive. For future developments.}

\item{maxEval}{Maximum fitness.}

\item{meanEval}{Average fitness.}
}
\description{
Function for graphically representing the evolution.
}
\examples{
\dontrun{
		#Graphical representation of the evolution after each generation.
		#Intended to be used by dGAselID() only.
		#Please refer to the example for dGAselID().
 }
}

