% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGAselID.R
\name{pointMutation}
\alias{pointMutation}
\title{pointMutation}
\usage{
pointMutation(individuals, mutationChance)
}
\arguments{
\item{individuals}{dataset returned by Individuals().}

\item{mutationChance}{chance for a point mutation to occur.}
}
\description{
Operator for the point mutation.
}
\examples{
\dontrun{
library(ALL)
data(ALL)

demoALL<-ALL[1:12,1:8]

set.seed(1234)
population<-InitialPopulation(demoALL, 4, 9)
individuals<-Individuals(population)

individuals
set.seed(123)
pointMutation(individuals, 4)
 }
}

