% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGAselID.R
\name{RandomAssortment}
\alias{RandomAssortment}
\title{RandomAssortment}
\usage{
RandomAssortment(newChrs, chrConf)
}
\arguments{
\item{newChrs}{Set of chromosomes.}

\item{chrConf}{Configuration of chromosomes returned by splitChromosomes().}
}
\description{
Random assortment of chromosomes operator.
}
\examples{
\dontrun{
library(ALL)
data(ALL)

demoALL<-ALL[1:12,1:8]

population02<-InitialPopulation(demoALL, 2, 4, FALSE)
chrConf02<-splitChromosomes(demoALL, 4)

set.seed(1357)
cr1<-Crossover(population02[1,], population02[2,], chrConf02)
RandomAssortment(cr1, chrConf02)
cr1
chrConf02
 }
}

