% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGAselID.R
\name{splitChromosomes}
\alias{splitChromosomes}
\title{splitChromosomes}
\usage{
splitChromosomes(x, noChr = 22)
}
\arguments{
\item{x}{Dataset in ExpressionSet format.}

\item{noChr}{Desired number of chromosomes. The default value is 22.}
}
\description{
Divides the genotypes into sets with a desired number of chromosomes.
}
\examples{
\dontrun{
library(ALL)
data(ALL)

demoALL<-ALL[1:12,1:8]

splitChromosomes(demoALL, 3)
splitChromosomes(demoALL)

 }
}

