% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGAselID.R
\name{dGAselID}
\alias{dGAselID}
\title{dGAselID}
\usage{
dGAselID(x, response, method = knn.cvI(k = 3, l = 2), trainTest = "LOG",
  startGenes, populationSize, iterations, noChr = 22, elitism = NA,
  ID = "ID1", pMutationChance = 0, nSMutationChance = 0,
  fSMutationChance = 0, lSDeletionChance = 0, wChrDeletionChance = 0,
  transposonChance = 0, randomAssortment = TRUE, embryonicSelection = NA,
  EveryGeneInInitialPopulation = TRUE, nnetSize = NA, nnetDecay = NA,
  rdaAlpha = NA, rdaDelta = NA, ...)
}
\arguments{
\item{x}{Dataset in ExpressionSet format.}

\item{response}{Response variable}

\item{method}{Supervised classifier for fitness evaluation. Most of the supervised classifiers in MLInterfaces are acceptable. The default is knn.cvI(k=3, l=2).}

\item{trainTest}{Cross-validation method. The default is "LOG".}

\item{startGenes}{Genes in the genotypes at initialization.}

\item{populationSize}{Number of genotypes in initial population.}

\item{iterations}{Number of iterations.}

\item{noChr}{Number of chromosomes. The default value is 22.}

\item{elitism}{Elite population in percentages.}

\item{ID}{Dominance. The default value is "ID1". Use "ID2" for Incomplete Dominance.}

\item{pMutationChance}{Chance for a Point Mutation to occur. The default value is 0.}

\item{nSMutationChance}{Chance for a Non-sense Mutation to occur. The default value is 0.}

\item{fSMutationChance}{Chance for a Frameshift Mutation to occur. The default value is 0.}

\item{lSDeletionChance}{Chance for a Large Segment Deletion to occur. The default value is 0.}

\item{wChrDeletionChance}{Chance for a Whole Chromosome Deletion to occur. The default value is 0.}

\item{transposonChance}{Chance for a Transposon Mutation to occur. The default value is 0.}

\item{randomAssortment}{Random Assortment of Chromosomes for recombinations. The default value is TRUE.}

\item{embryonicSelection}{Remove chromosomes with fitness < specified value. The default value is NA.}

\item{EveryGeneInInitialPopulation}{Request for every gene to be present in the initial population. The default value is TRUE.}

\item{nnetSize}{for nnetI. The default value is NA.}

\item{nnetDecay}{for nnetI. The default value is NA.}

\item{rdaAlpha}{for rdaI. The default value is NA.}

\item{rdaDelta}{for rdaI. The default value is NA.}

\item{...}{Additional arguments.}
}
\value{
The output is a list containing 5 named vectors, records of the evolution:
  \item{DGenes}{The occurrences in selected genotypes for every gene,}
  \item{dGenes}{The occurrences in discarded genotypes for every gene,}
  \item{MaximumAccuracy}{Maximum accuracy in every generation,}
  \item{MeanAccuracy}{Average accuracy in every generation,}
  \item{MinAccuracy}{Minimum accuracy in every generation,}
  \item{BestIndividuals}{Best individual in every generation.}
}
\description{
Initializes and starts the search with the genetic algorithm.
}
\examples{
\dontrun{
 library(genefilter)
 library(ALL)
 data(ALL)
 bALL = ALL[, substr(ALL$BT,1,1) == "B"]
 smallALL = bALL[, bALL$mol.biol \%in\% c("BCR/ABL", "NEG")]
 smallALL$mol.biol = factor(smallALL$mol.biol)
 smallALL$BT = factor(smallALL$BT)
 f1 <- pOverA(0.25, log2(100))
 f2 <- function(x) (IQR(x) > 0.5)
 f3 <- ttest(smallALL$mol.biol, p=0.1)
 ff <- filterfun(f1, f2, f3)
 selectedsmallALL <- genefilter(exprs(smallALL), ff)
 smallALL = smallALL[selectedsmallALL, ]
 rm(f1)
 rm(f2)
 rm(f3)
 rm(ff)
 rm(bALL)
 sum(selectedsmallALL)

 set.seed(149)
 res<-dGAselID(smallALL, "mol.biol", trainTest=1:79, startGenes=12, populationSize=200,
               iterations=150, noChr=5, pMutationChance=0.0075, elitism=4)
 }
}

