% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGAselID.R
\name{pointMutation}
\alias{pointMutation}
\title{pointMutation}
\usage{
pointMutation(individuals, mutationChance)
}
\arguments{
\item{individuals}{dataset returned by Individuals().}

\item{mutationChance}{chance for a point mutation to occur.}
}
\description{
Operator for the point mutation.
}
\examples{
\dontrun{
library(ALL)
data(ALL)

demoALL<-ALL[1:12,1:8]

set.seed(1357)
population01<-InitialPopulation(demoALL, 4, 4)
individuals01<-Individuals(population01)

individuals01
pointMutation(individuals01, 5)
 }
}

